add deploy script

+++ deploy.sh
@@ -1,21 +1,30 @@
-#!/bin/bash
+#!/bin/sh
 
-echo "start deploy to gh-pages"
+if [ "'git status -s'" ]
+then
+	echo "commit pending changes"
+	exit 1;
+fi
 
+echo "Deleting old public"
 rm -rf public
+mkdir public
+git worktree prune
+rm -rf .git/worktrees/public/
 
+echo "checking out gh-pages into public"
 git worktree add -B gh-pages public origin/gh-pages
 
+echo "Removing existing files"
+rm -rf public/*
+
+echo "generate site"
 hugo
 
-cd public 
-
+echo "Updating gh-pages branch"
+cd public
 git add --all
+git commit -m "publishing to gh-pages"
 
-git commit -m "Publishing to gh-pages"
-
-cd ..
-
-git push origin gh-pages
-
-echo "end deploy to gh-pages"
+#echo "pushing to github"
+#git push -all
